refactor(web): lift event bus + assistant lifecycle to RootLayout
The event bus was mounted at ChatLayout, which meant settings, logs,
and onboarding routes had no bus producer running — `useSettingsSync`
couldn't subscribe to `bus.sse.event` because nothing was opening the
SSE connection. The bus producer needs to be alive on every
authenticated route, not just chat.

RootLayout now:

- Calls `useAssistantLifecycle` once at the top of the authenticated
  tree and exposes the result via outlet context
  (`useRootOutletContext`).
- Calls `useEventBusInit` so the single SSE connection + DOM lifecycle
  listeners are alive on every child layout.

ChatLayout consumes `lifecycle` from the root outlet context instead
of running its own `useAssistantLifecycle` state machine. Its existing
`AssistantContextValue` outlet context for chat-scoped children is
unchanged.

Updates the convention notes (`apps/web/AGENTS.md`,
`docs/EVENT_BUS.md`, `assistant-context.ts` JSDoc) so future
contributors don't re-add lifecycle to ChatLayout.

The settings-sync migration from the previous commit now works: the
bus is alive on `/assistant/settings/*` because RootLayout — not
ChatLayout — owns the producer.
